+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/RandomPlaylistParametersPanel.java
@@ -144,7 +144,7 @@ final class RandomPlaylistParametersPanel extends JPanel {
 			if (noOfTracks.isNull()) {
 				return false;
 			}
-			if (genres.empty()) {
+			if (genres.isEmpty()) {
 				return false;
 			}